+++ cw-orch-daemon/src/channel.rs
@@ -3,7 +3,6 @@ use cosmrs::proto::cosmos::base::tendermint::v1beta1::{
 };
 use cw_orch_core::log::connectivity_target;
 use ibc_chain_registry::chain::Grpc;
-use ibc_relayer_types::core::ics24_host::identifier::ChainId;
 use tonic::transport::{Channel, ClientTlsConfig};
 
 use super::error::DaemonError;
@@ -13,7 +12,7 @@ pub struct GrpcChannel {}
 
 impl GrpcChannel {
     /// Connect to any of the provided gRPC endpoints
-    pub async fn connect(grpc: &[Grpc], chain_id: &ChainId) -> Result<Channel, DaemonError> {
+    pub async fn connect(grpc: &[Grpc], chain_id: &str) -> Result<Channel, DaemonError> {
         let mut successful_connections = vec![];
 
         for Grpc { address, .. } in grpc.iter() {
@@ -67,16 +66,14 @@ impl GrpcChannel {
                 .into_inner();
 
             // local juno does not return a proper ChainId with epoch format
-            if ChainId::is_epoch_format(&node_info.default_node_info.as_ref().unwrap().network) {
-                // verify we are connected to the spected network
-                if node_info.default_node_info.as_ref().unwrap().network != chain_id.as_str() {
-                    log::error!(
-                        "Network mismatch: connection:{} != config:{}",
-                        node_info.default_node_info.as_ref().unwrap().network,
-                        chain_id.as_str()
-                    );
-                    continue;
-                }
+            // verify we are connected to the expected network
+            if node_info.default_node_info.as_ref().unwrap().network != chain_id {
+                log::error!(
+                    "Network mismatch: connection:{} != config:{}",
+                    node_info.default_node_info.as_ref().unwrap().network,
+                    chain_id
+                );
+                continue;
             }
 
             // add endpoint to succesful connections