+++ packages/cw-orch-core/src/env.rs
@@ -3,16 +3,17 @@
 //! To get the env variable parsed value, you can use
 //! ```rust,no_run
 //! use cw_orch_core::CwOrchEnvVars;
-//! let env_variable = CwOrchEnvVars::load().unwrap().state_folder;
+//! let env_variable = CwOrchEnvVars::load().unwrap().state_file;
 //! ```
 
 use std::{env, path::PathBuf, str::FromStr};
+
+use cosmwasm_std::StdError;
 
 use crate::CwEnvError;
 
 const DEFAULT_TX_QUERY_RETRIES: usize = 50;
 
-pub const STATE_FOLDER_ENV_NAME: &str = "CW_ORCH_STATE_FOLDER";
 pub const STATE_FILE_ENV_NAME: &str = "STATE_FILE";
 pub const ARTIFACTS_DIR_ENV_NAME: &str = "ARTIFACTS_DIR";
 pub const GAS_BUFFER_ENV_NAME: &str = "CW_ORCH_GAS_BUFFER";
@@ -30,16 +31,13 @@ pub const TEST_MNEMONIC_ENV_NAME: &str = "TEST_MNEMONIC";
 pub const LOCAL_MNEMONIC_ENV_NAME: &str = "LOCAL_MNEMONIC";
 
 pub struct CwOrchEnvVars {
-    /// Optional - Absolute Path
-    /// Defaults to "~./cw-orchestrator"
-    /// This is the folder in which states of contracts are saved
-    /// This is not enforced to be an absolute path but this is highly recommended
-    pub state_folder: Option<PathBuf>,
-
     /// Optional - Path
-    /// /// This is the name of the state file
-    /// If the path is relative, this is taken from StateFolder
-    /// Defaults to "state.json"
+    /// This is the path to the state file
+    /// `folder/file.json` will resolve to `~/.cw-orchestrator/folder/file.json`
+    /// `./folder/file.json` will resolve `$pwd/folder/file.json`
+    /// `../folder/file.json` will resolve `$pwd/../folder/file.json`
+    /// `/usr/var/file.json` will resolve to `/usr/var/file.json`
+    /// Defaults to "~./cw-orchestrator/state.json"
     pub state_file: PathBuf,
 
     /// Optional - Path
@@ -109,11 +107,23 @@ pub struct CwOrchEnvVars {
     pub fee_granter: Option<String>,
 }
 
+/// Fetches the default state folder.
+/// This function should only error if the home_dir is not set and the `dirs` library is unable to fetch it
+/// This happens only in rare cases
+pub fn default_state_folder() -> Result<PathBuf, StdError> {
+    dirs::home_dir().map(|home| home.join(".cw-orchestrator"))
+        .ok_or( StdError::generic_err(
+            format!(
+                "Your machine doesn't have a home folder. You can't use relative path for the state file such as 'state.json'. 
+                Please use an absolute path ('/home/root/state.json') or a dot-prefixed-relative path ('./state.json') in the {} env variable.",
+                STATE_FILE_ENV_NAME
+            )))
+}
+
 impl Default for CwOrchEnvVars {
     fn default() -> Self {
         CwOrchEnvVars {
-            state_folder: dirs::home_dir().map(|home| home.join(".cw-orchestrator")),
-            state_file: PathBuf::from_str("state.json").unwrap(),
+            state_file: default_state_folder().unwrap().join("state.json"),
             artifacts_dir: None,
             gas_buffer: None,
             min_gas: None,
@@ -136,9 +146,6 @@ impl CwOrchEnvVars {
         let mut env_values = CwOrchEnvVars::default();
 
         // Then we load the values from env
-        if let Ok(str_value) = env::var(STATE_FOLDER_ENV_NAME) {
-            env_values.state_folder = Some(PathBuf::from_str(&str_value).unwrap());
-        }
         if let Ok(str_value) = env::var(STATE_FILE_ENV_NAME) {
             env_values.state_file = PathBuf::from_str(&str_value).unwrap();
         }