+++ packages/cw-orch-core/src/environment/mut_env.rs
@@ -5,7 +5,7 @@ use super::{
     queriers::bank::{BankQuerier, BankQuerierGetter},
     CwEnv, TxHandler,
 };
-use cosmwasm_std::{Addr, Coin};
+use cosmwasm_std::Coin;
 use cw_utils::NativeBalance;
 
 pub trait MutCwEnv: BankSetter + CwEnv {}
@@ -13,17 +13,18 @@ pub trait MutCwEnv: BankSetter + CwEnv {}
 pub trait BankSetter: TxHandler + BankQuerierGetter<Self::Error> {
     fn set_balance(
         &mut self,
-        address: &Addr,
+        address: impl Into<String>,
         amount: Vec<Coin>,
     ) -> Result<(), <Self as TxHandler>::Error>;
 
     fn add_balance(
         &mut self,
-        address: &Addr,
+        address: impl Into<String>,
         amount: Vec<Coin>,
     ) -> Result<(), <Self as TxHandler>::Error> {
+        let address = address.into();
         // Query the current balance of the account
-        let current_balance = self.bank_querier().balance(address, None)?;
+        let current_balance = self.bank_querier().balance(address.clone(), None)?;
         let future_balance = NativeBalance(current_balance) + NativeBalance(amount);
         // Set the balance with more funds
         self.set_balance(address, future_balance.into_vec())?;